requirements.txt
Django==5.1.7
djangorestframework==3.15.2
django-environ==0.11.2
google-generativeai==0.8.4
gunicorn==23.0.0
psycopg2-binary==2.9.10
pydub==0.25.1
requests==2.32.3
SpeechRecognition==3.14.1